feat(range-slider): replace implementation and set labels
